pyproject.toml
[tool.poetry]
name = "label-inspector"
version = "0.2.0"
readme = "README.md"
description = ""
authors = []
packages = [
    { include = "label_inspector" }
]
exclude = [
    "label_inspector/data/tests"
]


[tool.poetry.dependencies]
python = ">=3.9,<3.12"
ens-normalize = "^3.0.4"
fastapi = "^0.100.1"
hydra-core = "^1.3.2"
regex = "^2023.6.3"
unidecode = "^1.3.6"
emoji = "^2.7.0"
more-itertools = "^10.1.0"
mangum = {version = "^0.17.0", optional = true}


[tool.poetry.extras]
lambda = ["mangum"]


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-timeouts = "^1.2.1"
flake8 = "^6.1.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
httpx = "^0.24.1"
tqdm = "^4.65.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
